Update index.html
In the previous version, I was using the subreddit r/cat as the source for the cat pictures. However, I encountered some inappropriate pictures in that subreddit. To ensure a smoother and more enjoyable experience, I have now switched to using the Cat API. This API provides a faster and more reliable way to fetch cat pictures, ensuring that the displayed images are friendly and suitable for all viewers.